introduce unit squad class
