+++ Assets/Scripts/UnitSquad.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 26f45c39191c25643a7f13b2783c3f30
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
